requirements.txt
google-genai==1.20.0 
python-telegram-bot==22.1
groq==0.28.0
python-dotenv==1.0.1
watchdog==6.0.0
mistralai==1.8.2
boto3==1.35.81
nest_asyncio==1.6.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
pydantic==2.10.4
pydantic-settings==2.7.1